Move kmerge to their own file.
